backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pdfminer.six
fpdf2
python-dotenv
